requirements.txt
discord.py
psycopg2-binary